[None][perf] Fold the Q RoPE into q_b_layernorm for DSv4 context too
The generation half of this already landed. Context differs only in how a row's
position is found: query lengths are ragged, so `batch = token / seq_len` does not
hold. The kernel instead binary-searches `cu_q_seqlens` for the sequence owning a
token and computes `position = (token - seq_begin) + (kv_cache_len[b] -
current_seq_len)`; the second term is the chunked-prefill cached offset, matching
what `applyMLARopeAndAssignQKVKernelOptContext` derives. Every lane of a warp shares
its row, so the search is warp-uniform. Note `cu_q_seqlens` counts TOKENS here,
unlike the generation `seqQOffset`, which counts Q rows.

Once q_b_layernorm has rotated the rope segment into the FP8 Q buffer, the context
RoPE kernel must NOT run its Q region: the bf16 `q_pe` it would rotate is left stale
by the fused kernel, so rotating it would overwrite good FP8 rope slots with
garbage. `MlaParams::q_rope_done` carries that, threaded from `mla.py` through
`AttentionForwardArgs` -> nanobind -> `thop::attention` -> `invokeMLARopeContext`,
and the Q region early-returns. The return sits after the `kOutputFp8Q` bmm-scale
prologue, which runs on block (0,0,0) outside that branch, so the scheduler scales
are still emitted.

The token-wise cumsum comes from a new `mla_prepare_ctx_cu_seqlens`, built once per
iteration into a fixed-address buffer. It deliberately does not reuse
`ctx_uncached_token_indptr`, which holds the same values but only exists when
`enable_context_mla_with_cached_kv` is set: with chunked prefill and block reuse
both off, that buffer is absent and the fusion silently fell back to the old path --
GSM8K still passed, and only the kernel trace showed it had never engaged.

Measured on GB200, CTX rank 0, ISL 8192, one build, arms differing only by
`TRTLLM_DISABLE_FUSED_Q_ROPE`:

  applyMLARopeAndAssignQKVKernelOptContext  70443 -> 23565 ns  (-66.5%)
  deepseekV4QNormFusedKernel               205310 -> 225194 ns  (+9.7%)
  combined                                 275752 -> 248759 ns  (-9.8%)
  share of CTX GPU time                       2.7% -> 2.4%

Prefill is MoE-communication bound, so this is ~0.3pp of context GPU time, not a
throughput win; it is worth having for symmetry with the generation path and for
the launch it removes. GSM8K full accuracy 96.399.

(cherry picked from commit 270c968ba0d8777bd5b4e1f4a486dbd164f26651)

+++ cpp/tensorrt_llm/kernels/deepseekV4QNormKernel.cu
@@ -164,7 +164,7 @@ __global__ void deepseekV4QNormFusedKernel(T const* __restrict__ input, __nv_fp8
     T* __restrict__ q_pe_out, float const* __restrict__ quant_scale_qkv_ptr, int totalRows,
     int quantQNopeRowStrideBytes, float eps, float2 const* __restrict__ cos_sin_cache = nullptr,
     int const* __restrict__ cache_seq_lens = nullptr, int num_heads = 0, int seq_len = 0,
-    int64_t const* __restrict__ cu_q_seqlens = nullptr, int num_seqs = 0)
+    int const* __restrict__ cu_q_seqlens = nullptr, int num_seqs = 0)
 {
     static_assert(kHeadDim % (2 * kWarpSize) == 0);
     static_assert(kNopeDim > 0 && kNopeDim < kHeadDim);
@@ -250,8 +250,8 @@ __global__ void deepseekV4QNormFusedKernel(T const* __restrict__ input, __nv_fp8
                         hi = mid - 1;
                     }
                 }
-                int const seqBegin = static_cast<int>(cu_q_seqlens[lo]);
-                int const currentSeqLen = static_cast<int>(cu_q_seqlens[lo + 1]) - seqBegin;
+                int const seqBegin = cu_q_seqlens[lo];
+                int const currentSeqLen = cu_q_seqlens[lo + 1] - seqBegin;
                 positionId = (token - seqBegin) + (cache_seq_lens[lo] - currentSeqLen);
             }
             else
@@ -281,7 +281,7 @@ __global__ void deepseekV4QNormFusedKernel(T const* __restrict__ input, __nv_fp8
 template <typename T>
 void dispatchDeepseekV4QNormFused(void const* input, void* quant_q_nope, void* q_pe_out,
     void const* quant_scale_qkv_ptr, int totalRows, int headDim, int nopeDim, int quantQNopeRowStrideBytes, float eps,
-    void const* cos_sin_cache, int const* cache_seq_lens, int num_heads, int seq_len, int64_t const* cu_q_seqlens,
+    void const* cos_sin_cache, int const* cache_seq_lens, int num_heads, int seq_len, int const* cu_q_seqlens,
     int num_seqs, cudaStream_t stream)
 {
     assert(headDim == 512);
@@ -336,7 +336,7 @@ void invokeDeepseekV4QNorm(
 void invokeDeepseekV4QNormFusedFp8(void const* input, void* quant_q_nope, void* q_pe_out,
     void const* quant_scale_qkv_ptr, int totalRows, int headDim, int nopeDim, int quantQNopeRowStrideBytes,
     bool isBfloat16, float eps, void const* cos_sin_cache, int const* cache_seq_lens, int num_heads, int seq_len,
-    int64_t const* cu_q_seqlens, int num_seqs, cudaStream_t stream)
+    int const* cu_q_seqlens, int num_seqs, cudaStream_t stream)
 {
     if (totalRows == 0)
     {
