+++ cpp/tensorrt_llm/kernels/deepseekV4QNormKernel.h
@@ -53,7 +53,7 @@ void invokeDeepseekV4QNorm(
 void invokeDeepseekV4QNormFusedFp8(void const* input, void* quant_q_nope, void* q_pe_out,
     void const* quant_scale_qkv_ptr, int totalRows, int headDim, int nopeDim, int quantQNopeRowStrideBytes,
     bool isBfloat16, float eps, void const* cos_sin_cache, int const* cache_seq_lens, int num_heads, int seq_len,
-    int64_t const* cu_q_seqlens, int num_seqs, cudaStream_t stream);
+    int const* cu_q_seqlens, int num_seqs, cudaStream_t stream);
 
 } // namespace kernels
